Device-first OS list filtering
Rather than a drop down dialog, which could present users with images
that may not run on their hardware, allow selection of Raspberry Pi as a
first stage. If users adopt this feature, they are presented with a
subset of images that we know will actually run on their hardware.

This is achieved by leveraging @maxnet's excellent OS filtering scheme.

Future work will attach image and description support to this OS list.