qml: Enforce maximum Pi logo size
The source image is a PNG measuring 626x418. Prior to this change,
it was possible to resize the window such that window.height/2 evaluated
to much larger than 418dips - distorting the image.

Rejig how we calculate the size of the red rectangle and the image
rectangle, such that the logo image cannot exceed 418dips of height.